requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==22.1
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
phonenumbers==8.12.53
Pillow==9.2.0
pytz==2022.2
sqlparse==0.4.2
